refactor(proxy): internalize sse translation and cleanup handlers
